requirements.txt
aiohttp
pandas
python-dotenv
requests
tqdm